reviewer update and changing other Lit shader passes to have same changes as it reproduces as well

+++ com.unity.render-pipelines.lightweight/Shaders/Particles/ParticlesLitForwardPass.hlsl
@@ -22,16 +22,16 @@ struct VaryingsParticle
 {
     half4 color                     : COLOR;
     float2 texcoord                 : TEXCOORD0;
-    
+
     float4 positionWS               : TEXCOORD1;
 
 #ifdef _NORMALMAP
-    half4 normalWS                  : TEXCOORD2;    // xyz: normal, w: viewDir.x
-    half4 tangentWS                 : TEXCOORD3;    // xyz: tangent, w: viewDir.y
-    half4 bitangentWS               : TEXCOORD4;    // xyz: bitangent, w: viewDir.z
+    float4 normalWS                 : TEXCOORD2;    // xyz: normal, w: viewDir.x
+    float4 tangentWS                : TEXCOORD3;    // xyz: tangent, w: viewDir.y
+    float4 bitangentWS              : TEXCOORD4;    // xyz: bitangent, w: viewDir.z
 #else
-    half3 normalWS                  : TEXCOORD2;
-    half3 viewDirWS                 : TEXCOORD3;
+    float3 normalWS                 : TEXCOORD2;
+    float3 viewDirWS                : TEXCOORD3;
 #endif
 
 #if defined(_FLIPBOOKBLENDING_ON)
@@ -67,13 +67,13 @@ void InitializeInputData(VaryingsParticle input, half3 normalTS, out InputData o
 #endif
 
     output.normalWS = NormalizeNormalPerPixel(output.normalWS);
-    
+
 #if SHADER_HINT_NICE_QUALITY
     viewDirWS = SafeNormalize(viewDirWS);
 #endif
 
     output.viewDirectionWS = viewDirWS;
-    
+
 #if defined(_MAIN_LIGHT_SHADOWS) && !defined(_RECEIVE_SHADOWS_OFF)
     output.shadowCoord = input.shadowCoord;
 #else
@@ -92,15 +92,15 @@ void InitializeInputData(VaryingsParticle input, half3 normalTS, out InputData o
 VaryingsParticle ParticlesLitVertex(AttributesParticle input)
 {
     VaryingsParticle output = (VaryingsParticle)0;
-    
+
     UNITY_SETUP_INSTANCE_ID(input);
     UNITY_TRANSFER_INSTANCE_ID(input, output);
     UNITY_INITIALIZE_VERTEX_OUTPUT_STEREO(output);
-    
+
     VertexPositionInputs vertexInput = GetVertexPositionInputs(input.vertex.xyz);
     VertexNormalInputs normalInput = GetVertexNormalInputs(input.normal, input.tangent);
     half3 viewDirWS = GetCameraPositionWS() - vertexInput.positionWS;
-    
+
 #if !SHADER_HINT_NICE_QUALITY
     viewDirWS = SafeNormalize(viewDirWS);
 #endif
@@ -123,13 +123,13 @@ VaryingsParticle ParticlesLitVertex(AttributesParticle input)
     output.positionWS.w = fogFactor;
     output.clipPos = vertexInput.positionCS;
     output.color = input.color;
-    
+
     output.texcoord = input.texcoords.xy;
 #ifdef _FLIPBOOKBLENDING_ON
     output.texcoord2AndBlend.xy = input.texcoords.zw;
     output.texcoord2AndBlend.z = input.texcoordBlend;
 #endif
-    
+
 #if defined(_SOFTPARTICLES_ON) || defined(_FADING_ON) || defined(_DISTORTION_ON)
     output.projectedPosition = vertexInput.positionNDC;
 #endif
@@ -144,8 +144,8 @@ VaryingsParticle ParticlesLitVertex(AttributesParticle input)
 half4 ParticlesLitFragment(VaryingsParticle input) : SV_Target
 {
     UNITY_SETUP_INSTANCE_ID(input);
-    UNITY_SETUP_STEREO_EYE_INDEX_POST_VERTEX(input);    
-    
+    UNITY_SETUP_STEREO_EYE_INDEX_POST_VERTEX(input);
+
     float3 blendUv = float3(0, 0, 0);
 #if defined(_FLIPBOOKBLENDING_ON)
     blendUv = input.texcoord2AndBlend;
@@ -158,13 +158,13 @@ half4 ParticlesLitFragment(VaryingsParticle input) : SV_Target
 
     SurfaceData surfaceData;
     InitializeParticleLitSurfaceData(input.texcoord, blendUv, input.color, projectedPosition, surfaceData);
-    
+
     InputData inputData = (InputData)0;
     InitializeInputData(input, surfaceData.normalTS, inputData);
 
     half4 color = LightweightFragmentPBR(inputData, surfaceData.albedo,
-        surfaceData.metallic, half3(0, 0, 0), surfaceData.smoothness, surfaceData.occlusion, surfaceData.emission, surfaceData.alpha);
-    
+        surfaceData.metallic, /* specular */ half3(0, 0, 0), surfaceData.smoothness, surfaceData.occlusion, surfaceData.emission, surfaceData.alpha);
+
     color.rgb = MixFog(color.rgb, inputData.fogCoord);
     return color;
 }
