+++ com.unity.render-pipelines.lightweight/Shaders/Particles/ParticlesSimpleLitForwardPass.hlsl
@@ -23,16 +23,16 @@ struct VaryingsParticle
 {
     half4 color                     : COLOR;
     float2 texcoord                 : TEXCOORD0;
-    
+
     float4 positionWS               : TEXCOORD1;
 
 #ifdef _NORMALMAP
-    half4 normalWS                  : TEXCOORD2;    // xyz: normal, w: viewDir.x
-    half4 tangentWS                 : TEXCOORD3;    // xyz: tangent, w: viewDir.y
-    half4 bitangentWS               : TEXCOORD4;    // xyz: bitangent, w: viewDir.z
+    float4 normalWS                 : TEXCOORD2;    // xyz: normal, w: viewDir.x
+    float4 tangentWS                : TEXCOORD3;    // xyz: tangent, w: viewDir.y
+    float4 bitangentWS              : TEXCOORD4;    // xyz: bitangent, w: viewDir.z
 #else
-    half3 normalWS                  : TEXCOORD2;
-    half3 viewDirWS                 : TEXCOORD3;
+    float3 normalWS                 : TEXCOORD2;
+    float3 viewDirWS                : TEXCOORD3;
 #endif
 
 #if defined(_FLIPBOOKBLENDING_ON)
@@ -68,13 +68,13 @@ void InitializeInputData(VaryingsParticle input, half3 normalTS, out InputData o
 #endif
 
     output.normalWS = NormalizeNormalPerPixel(output.normalWS);
-    
+
 #if SHADER_HINT_NICE_QUALITY
     viewDirWS = SafeNormalize(viewDirWS);
 #endif
 
     output.viewDirectionWS = viewDirWS;
-    
+
 #if defined(_MAIN_LIGHT_SHADOWS) && !defined(_RECEIVE_SHADOWS_OFF)
     output.shadowCoord = input.shadowCoord;
 #else
@@ -93,7 +93,7 @@ void InitializeInputData(VaryingsParticle input, half3 normalTS, out InputData o
 VaryingsParticle ParticlesLitVertex(AttributesParticle input)
 {
     VaryingsParticle output;
-    
+
     UNITY_SETUP_INSTANCE_ID(input);
     UNITY_TRANSFER_INSTANCE_ID(input, output);
     UNITY_INITIALIZE_VERTEX_OUTPUT_STEREO(output);
@@ -120,7 +120,7 @@ VaryingsParticle ParticlesLitVertex(AttributesParticle input)
     output.positionWS.w = ComputeFogFactor(vertexInput.positionCS.z);
     output.clipPos = vertexInput.positionCS;
     output.color = input.color;
-    
+
     output.texcoord = input.texcoords.xy;
 #ifdef _FLIPBOOKBLENDING_ON
     output.texcoord2AndBlend.xy = input.texcoords.zw;
@@ -142,7 +142,7 @@ half4 ParticlesLitFragment(VaryingsParticle input) : SV_Target
 {
     UNITY_SETUP_INSTANCE_ID(input);
     UNITY_SETUP_STEREO_EYE_INDEX_POST_VERTEX(input);
-    
+
     float2 uv = input.texcoord;
     float3 blendUv = float3(0, 0, 0);
 #if defined(_FLIPBOOKBLENDING_ON)
@@ -165,7 +165,7 @@ half4 ParticlesLitFragment(VaryingsParticle input) : SV_Target
 #endif
     half4 specularGloss = SampleSpecularSmoothness(uv, blendUv, albedo.a, _SpecColor, TEXTURE2D_ARGS(_SpecGlossMap, sampler_SpecGlossMap));
     half shininess = specularGloss.a;
-    
+
 #if defined(_DISTORTION_ON)
     diffuse = Distortion(half4(diffuse, alpha), normalTS, _DistortionStrengthScaled, _DistortionBlend, projectedPosition);
 #endif
